Renamed firestore.ts to transaction.ts
